Allow all subdomains again
I *think* the problem before was that this *was* actually working, but I
also need to explicitly check the `origin` and `referer` headers (which
I will do in a follow-on change). But I need to confirm that, and the
best way to do so *right now* is simply to deploy this thing. Ha!