lib: support multithreading and only load fonts once
I am not currently using this in any direct sense, but if at some point
I decide I want to generate these and cache them as part of the build,
this is handy, and it was useful for testing throughput in the CLI. I
can render 1,000 images in about 2.5 seconds, which is (a) great and (b)
way slower than I want for the actual build ever to take. Also, that's
on this super fast M4 Max MBP; it would be *much* slower on e.g. Render.